feat: implementação de regra para editar empresa e filial
feat: implementação de regra para editar empresa e filial

+++ 2 - Application/1-Services/CompanyService/CompanyService.cs
@@ -62,19 +62,20 @@ public class CompanyService : BaseService
             if (user == null)
                 return ErrorResponse(UserLoginMessage.InvalidCredentials);
 
-            // Obtém as empresas relacionadas ao usuário
+            // Verifica se o usuário tem permissão para atualizar a empresa
+            var hasPermission = await _companyRepository.ExistsEditCompanyUser(dto.UserId, id);
+            if (!hasPermission)
+                return ErrorResponse(Message.Unauthorized);
+
+            // Obtém a empresa relacionada ao usuário
             var companies = await _companyRepository.GetCompaniesByUserId(dto.UserId);
+            var company = companies.FirstOrDefault(a => a.Id == id);
 
-            var company = companies.Where(a => a.Id == id).FirstOrDefault();
             if (company == null)
-            {
                 return ErrorResponse(Message.NotFound);
-            }
-            else
-            {
-                company.Name = dto.Name;
-            }
 
+            // Atualiza os dados da empresa
+            company.Name = dto.Name;
             await _companyRepository.UpdateCompany(company);
 
             return SuccessResponse(Message.Success);
@@ -85,6 +86,7 @@ public class CompanyService : BaseService
         }
     }
 
+
     public async Task<ResultValue> DeleteCompany(int userId, int id)
     {
         try
@@ -93,7 +95,14 @@ public class CompanyService : BaseService
             if (user == null)
                 return ErrorResponse(UserLoginMessage.InvalidCredentials);
 
+            // Verifica se o usuário tem permissão para excluir a empresa
+            var hasPermission = await _companyRepository.ExistsCompanyUser(userId, id);
+            if (!hasPermission)
+                return ErrorResponse(Message.Unauthorized);
+
             var company = await _companyRepository.GetById(id);
+            if (company == null)
+                return ErrorResponse(Message.NotFound);
 
             await _companyRepository.DeleteCompany(company.Id);
 
@@ -192,7 +201,7 @@ public class CompanyService : BaseService
         }
     }
 
-    public async Task<ResultValue> UpdateSubCompany(int id, UpdateCompanyDto dto)
+    public async Task<ResultValue> UpdateSubCompany(int id, UpdateSubCompanyDto dto)
     {
         try
         {
@@ -200,20 +209,19 @@ public class CompanyService : BaseService
             if (user == null)
                 return ErrorResponse(UserLoginMessage.InvalidCredentials);
 
-            // Obtém as empresas relacionadas ao usuário
-            var companies = await _companyRepository.GetSubCompanieByUserId(dto.UserId);
+            // Verifica se o usuário tem permissão para editar a subempresa
+            var hasPermission = await _companyRepository.ExistsEditSubCompanyUser(dto.UserId, dto.CompanyId, id);
+            if (!hasPermission)
+                return ErrorResponse(Message.Unauthorized);
 
-            var subCompany = companies.SubCompanies.Where(a => a.Id == id).FirstOrDefault();
-
+            // Obtém a subempresa relacionada ao usuário
+            var subCompanies = await _companyRepository.GetSubCompanieByUserId(dto.UserId);
+            var subCompany = subCompanies.SubCompanies.FirstOrDefault(a => a.Id == id);
 
             if (subCompany == null)
-            {
                 return ErrorResponse(Message.NotFound);
-            }
-            else
-            {
-                subCompany.Name = dto.Name;
-            }
+
+            subCompany.Name = dto.Name;
 
             await _companyRepository.UpdateSubCompany(subCompany);
 
@@ -225,7 +233,8 @@ public class CompanyService : BaseService
         }
     }
 
-    public async Task<ResultValue> DeleteSubCompany(int userId, int companyId, int subcompanyId)
+
+    public async Task<ResultValue> DeleteSubCompany(int userId, int companyId, int subCompanyId)
     {
         try
         {
@@ -233,22 +242,20 @@ public class CompanyService : BaseService
             if (user == null)
                 return ErrorResponse(UserLoginMessage.InvalidCredentials);
 
+            // Verifica se o usuário tem permissão para excluir a subempresa
+            var hasPermission = await _companyRepository.ExistsEditSubCompanyUser(userId, companyId, subCompanyId);
+            if (!hasPermission)
+                return ErrorResponse(Message.Unauthorized);
+
             var company = await _companyRepository.GetById(companyId);
-
             if (company == null)
-            {
                 return ErrorResponse(Message.NotFound);
-            }
 
-            var subcompany = company.SubCompanies.Where(a => a.Id == subcompanyId).FirstOrDefault();
-
-            if (subcompany == null)
-            {
+            var subCompany = company.SubCompanies.FirstOrDefault(a => a.Id == subCompanyId);
+            if (subCompany == null)
                 return ErrorResponse(Message.NotFound);
-            }
 
-
-            await _companyRepository.DeleteSubCompany(company.Id, subcompany.Id);
+            await _companyRepository.DeleteSubCompany(company.Id, subCompany.Id);
 
             return SuccessResponse(Message.DeleteSuccess);
         }
@@ -257,6 +264,7 @@ public class CompanyService : BaseService
             return ErrorResponse(ex);
         }
     }
+
 
     public async Task<ResultValue> CreateUserSubCompany(CreateSubCompanyUserDto dto)
     {
@@ -374,40 +382,60 @@ public class CompanyService : BaseService
     {
         try
         {
-
+            // Verifica se o usuário existe
             var user = await _userRepository.GetByUserId(userId);
             if (user == null)
                 return ErrorResponse(UserLoginMessage.InvalidCredentials);
 
-
+            // Obtém as empresas paginadas
             var companies = await _companyRepository.GetCompaniesByUserIdPaginated(userId, skip, take);
-
             if (companies == null || !companies.Any())
                 return ErrorResponse(Message.NotFound);
 
-
-            var companyDtos = companies.Select(c => new CompanyDto
+            // Mapeia as empresas e permissões para o response
+            var companyResponses = companies.Select(company => new CompanyUsersimpleDto
             {
-                Id = c.Id,
-                Name = c.Name,
-                DateCreate = c.DateCreate,
-                SubCompanies = c.SubCompanies?.Select(sub => new SubCompanyDto
+                CompanyId = company.Id,
+                CompanyName = company.Name,
+                DateCreate = company.DateCreate,
+                Permission = company.CompanyUsers?.FirstOrDefault()?.Permission != null ? new PermissionResponse
                 {
-                    Id = sub.Id,
-                    Name = sub.Name,
-                    DateCreate = sub.DateCreate,
-                    CompanyId = c.Id,
-                    CompanyName = c.Name
-                }).ToList() ?? new List<SubCompanyDto>()
+                    Id = company.CompanyUsers.FirstOrDefault().Permission.Id,
+                    Name = company.CompanyUsers.FirstOrDefault().Permission.Name
+                } : null,
+                SubCompanies = company.SubCompanies?
+                    .Where(subCompany => subCompany.CompanyUsers.Any(cu => cu.UserId == userId)) // Filtra apenas as vinculadas ao usuário
+                    .Select(subCompany => new SubCompanyUsersimpleDto
+                    {
+                        SubCompanyId = subCompany.Id,
+                        SubCompanyName = subCompany.Name,
+                        CompanyId = company.Id,
+                        DateCreate = subCompany.DateCreate,
+                        Permission = subCompany.CompanyUsers?
+                            .FirstOrDefault(cu => cu.UserId == userId)?.Permission != null ? new PermissionResponse
+                            {
+                                Id = subCompany.CompanyUsers.FirstOrDefault(cu => cu.UserId == userId).Permission.Id,
+                                Name = subCompany.CompanyUsers.FirstOrDefault(cu => cu.UserId == userId).Permission.Name
+                            } : null
+                    }).ToList() ?? new List<SubCompanyUsersimpleDto>()
             }).ToList();
 
-            return SuccessResponse(companyDtos);
+            // Resposta final com paginação
+            var response = new CompanyUserDto
+            {
+                UserId = userId,
+                Name = user.Name,
+                companies = companyResponses ?? new List<CompanyUsersimpleDto>()
+            };
+
+            return SuccessResponse(response);
         }
         catch (Exception ex)
         {
             return ErrorResponse(ex);
         }
     }
+
 
     #endregion
 }
